GOCBC-1232: Only call observer End if req not already completed
Motivation
----------
When we receive a stream end dcp packet then we call the relevant
user callback before calling internalCancel to cancel the request.
It's possble that at the same time we receive a config indicating
that the node owning that stream has been rebalanced out which
leads to a graceful close of the memdclient, which forcefully
cancels any persistent requests. This can lead to the event handler
End function being called a second time. We should mark the request
as completed before calling the oberver End callback.

Changes
--------
Complete the request before calling the stream observer End callback.
Before calling the End callback also check that the internalCancel
function succeeds.

Change-Id: I1852b44565aa7b11a50d87b46db4539b218d4a1a
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/169795
Reviewed-by: James Lee <james.lee@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>