GOCBC-1233: Ensure bootstrap request sends succeed before creating timers
Motivation
----------
If a pipelineclient closes its cancel dial signal whilst a sasl
auth request is in the process of being sent on a closed memdclient
then we can end up in a position where we attempt to write to the
kvmux auth handler's completed channel (buffer size of 1) twice.
This can lead to agent close hanging unable to complete. This
happens because we create the request timer and goroutine to handle
success/cancellation before the request has been sent.

Changes
-------
Move timer and handler goroutine creation to after bootstrap
requests have been successfully dispatched.

Change-Id: I414f08ec88bb4c9325aa1c2852b689f3e03023d5
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/169832
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>