TXNG-165: Fail cleanup attempt on first error
Motivation
-----------
At the moment we almost follow the spec, we do cleanup of
committed inserts, committed replaces, rolled back inserts
etc.. in serial and we return any errors as soon as we
encounter them. These are batches of docs, we cleanup each
doc in serial but we do not return any error as soon as we
encounter it. We cleanup all of the docs in the batch and
then return the first error we encountered, if any.

This means that cleanup threads that are racing can
effectively end up interleaving with each other and cause
cas mismatches on different docs causing all of them to then
fail.

Changes
-------
Return error from cleanup attempt as soon as we encounter
one.

Change-Id: Ibb503816a01e3e72f41fb5a5d92224f336236f36
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/170704
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>