GOCBC-1242: Flush the DCP buffer on EOF originating from remote
Motivation
----------
When we detect an EOF on a connection when we go through all
outstanding requests and call their callbacks with EOF.

We should change this so that the DCP queue is flushed first
(which means that any outstanding End messages will be handled
and an EOF not returned), but only when the EOF is not due to
the connection being closed our side.

Changes
-------
Move shuttingDown flag to be owned by the memdclient and only set
when Close is called.

Change-Id: I0d226c8427b2f4911de9ab7462220b6ff944182b
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/170813
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>