GOCBC-1242: Flush the DCP buffer on EOF originating from remote
Motivation
----------
When we detect an EOF on a connection when we go through all
outstanding requests and call their callbacks with EOF.

We should change this so that the DCP queue is flushed first
(which means that any outstanding End messages will be handled
and an EOF not returned), but only when the EOF is not due to
the connection being closed our side.

Changes
-------
Move shuttingDown flag to be owned by the memdclient and only set
when Close is called.

Change-Id: I921bfaba3dd44d97e9074853fe838552dde7da83
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/170880
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>