GOCBC-1248: Check if client already closed during graceful shutdown
Motivation
-----------
The fix for GOCBC-1242 introduced a change to graceful shutdown
where the client calls straight into closeConn rather than
Close when the op list becomes empty. This means that if the
memdclient closes (e.g. a read failure on the conn) during
graceful shutdown then we don't check if the client is already
closed from the graceful shutdown side and we try to close it
twice, which leads to a panic.

Changes
-------
When op list is empty and the memdclient is in graceful shutdown
then call Close rather than closeConn.
Add a nil check guard to memdconn Release function.

Change-Id: I7e41e1b8aedb1443ab655863cc502e480d7e5cb0
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/171513
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>