GOCBC-1256: Only fallback to http polling if there are http addresses
Motivation
----------
When bootstrap fails on select bucket with a enoent error
then we fallback to http polling to fetch a cluster config (as
kv connections will never fully build so cccp will never run).
This can happen even when there are no known http addresses to
poll against. When this happens it's possible for the bucket to
come online and connections successfully get built, however the
SDK will never fetch a config because the http poller will be
sat doing nothing.

Changes
-------
Only allow http fallback when the http muxer has management endpoints
to query against.

Change-Id: I6ea57a58d144ac80d5917e55cabeff2d0df9d1ac
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/171869
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>