GOCBC-1265: Bundle Capella CA certificates with the SDK
Motivation
----------
Streamline configuring the SDK to connect to a Capella cluster,
so users don't need to download the CA certificate from the
Capella website.

Changes
-------------
Bundle the Capella CA certificate with the SDK. If the user does
not configure a different ca provider, trust the Capella CA.

Return an error if the user doesn't specify a ca provider
when connecting to a non-Capella cluster using TLS.

If the user tries connecting to Capella without TLS, log a warning
and let the attempt continue.

Change-Id: I3fbb8cb575087dedd55d4250bd8a957b0947a066
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/172936
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>