GOCBC-1159: Modifications to improve support for serverless envs
Motivation
-----------
At the moment we do not officially support serverless environments
like aws lambda. We would like to support these environments so we
should improve our resilience to the sorts of errors that these
environments can cause.

Changes
-------
Add HTTP connect timeout config option so that dial timeouts can
be set by the user.
Update HTTP component to attempt retry with socket not available
reason when a request fails due to dial timeout or connection
refused.
Update kv mux to attempt retry with socket not available when
the muxer is not closed and the request errors with EOF and has
definitely not been written to the network or the memdclient owning
the connection has closed itself.
A few QOL improvements to logging.
Change-Id: I8b202d15f8783e829eb5aa6e224c01d6e0ba8593
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/175342
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>