GOCBC-1297: Add finalErrorToRaise to txn state bits
Motivation
----------
As a part of ExtThreadSafety a finalErrorToRaise has been added
to statebits. Changing to this will allow gocb to simplify its
core loop and bring us closer in line with specs.

Changes
-------
Add finalErrorToRaise to txn state bits.
Change Get to not include a toRaise reason when the error is
DocumentNotFound.
Change rollback to never set toRaise.
Change-Id: I6a9df77a34444a3e6b2b390fcd90eef9249d9c0b
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/175555
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>