GOCBC-1298: Mask error causes in TransactionOperationFailed
Motivation
----------
When a user uses Unwrap on a TOF error we return the underlying
error cause. TOF is intended to be opaque so we should actually
just return a "transaction operation failed" error when Unwrap
is used. In order for gocb to extract the actual error we should
add an InternalUnwrap function. We cannot change the JSON format
of the TOF error, this must remain the same.

Changes
-------
Create a sentinel errTransactionOperationFailed error and return
it when Unwrap is called on TOF.
Add an InternalUnwrap function to TOF which returns the actual
underlying error cause.
Directly return ErrDocumentNotFound from Get where applicable.

Change-Id: Ia9fa083fada90432fb3b741650f04fac607ef5af
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/175742
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>