GOCBC-1300: Check ProcessATR in lost cleanup for collection unknown
Motivation
-----------
When lost cleanup fails to process the client record or an atr
with a timeout error containing a collection unknown reason then
we should be removing that collection from our cleanup locations.
This is not happening correctly - we ignore all errors from
processing atrs.

Changes
-------
Move collection unknown detection to happen after processing
the client record and after processing each atr.
When a collection unknown is detected return a collection
not found upstream to trigger the location removal.

Change-Id: I3bb1f364e6de9e3224796456d54d5a4cb1f13f9c
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/175879
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>