GOCBC-1301: Set addLostCleanupLocation in ResumeTransactionAttempt
Motivation
----------
When we added the feature to automatically add new metadata
locations to cleanup a code path was missed to set this function
on a resumed attempt. This means that attempts that have been
serialized and then resumed lead to a panic when serialization
occurs before any operations take place.

Changes
-------
Set addLostCleanupLocation on the attempt when an attempt is
resumed using ResumeTransactionAttempt.

Change-Id: I4a201974d22b6b56ba0cd8bdcafce0a857358f74
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/176237
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>