GOCBC-1284: Remove init fn from bootstrapping
Motivation
----------
At the moment we pass a callback into bootstrap which is called
at the end of the bootstrap process to perform any agent level
calls, i.e. DCP controls. We should create a dcp bootstrap client
which can do these controls itself as a part of bootstrap.

Changes
-------
Add a dcp boostrap client which wraps the bootstrap client and
exposes DCP control functions, in a synchronous fashion.
Update dialler to perform standard bootstrap or dcp bootstrap
depending on whether the DCP agent is being used.
Remove the concept of the memdInitFunc.

Change-Id: Ib6baf81e02d08f9b24e2321c1e583b851753850f
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/177087
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>