GOCBC-1321: Don't always wait for services to be available
Motivation
-----------
When the user does not specify any services and does an AgentGroup/GCCCP
level waituntilready then we currently default to waiting for all
HTTP services.

Now that we have updated WaitUntilReady to not consider a missing
service as ready this means that it can fail in unexpected ways.
We need to configure the diagnostics component so that it will
only force a wait for a missing service when that service has
been explicitly asked for by the user.

Changes
-------
Update the diagnostics component so that its internal waituntilready
call takes a forceWait parameter, which forces the SDK to wait until
desired services to come online.
Move default service types out to caller level - agent and dcp agent.
If the user does not specify any service types then do not force wait.
Change-Id: I4b533cd6121f4ab383f302413b6b8fbd76f958fe
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/177501
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>