GOCBC-1331: Pipeline fetching config fetch into initial bootsrap
Motivation
----------
In order to support further evolution of our bootstrap process
we need to be able to fetch (and pipeline) a config fetch into
the initial bootstrap of connections. This will allow us to remove
the fallback from select bucket and, for now, replace it with a
fallback based on whether fetching a config succeeds.

Changes
-------
Add the ability to fetch a config to the bootstrap process.
Update config manager to use locks as configs can now come multiple
places concurrently (note that this was actually already possible due
to NMV).
Add a run loop to poller controller so that we will bounce between
cccp and http polling as needed.
Update the poller controller so that switching poller is always done
in the run loop.

Change-Id: Icaf84eb6ea683b7a07d3066ce6eac56aa65bc391
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/178688
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>