GOCBC-1335: Check for ValueTooLarge during txn error classification
Motivation
----------
This shouldn't really be needed as ErrValueTooLarge is effectively
the same error as ErrMemdTooBig, however the FIT hooks are used
to simulate an ATR becoming full. From FIT we can only return a
ErrValueTooLarge as gocb does not have any errors which translate
the Memd class of errors.

Changes
-------
Add a check for ErrValueTooLarge in txn error classification.

Change-Id: Ib6171c58a4054d2d4ab682c36e7b92c1a26febae
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/178746
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>