GOCBC-1293: Add support for resource units
Motivation
----------
We need the ability to surface the number of resource units
that an operation has used.

Changes
-------
Add a resource units type.
Add resource units to all kv result types.
Add read and write resource unit frames to memd and to packet.
Read any resource units into the packet and populate the values
on the result types.

Change-Id: I655ecb6275ff46b3a3994ea527d10fa34b5592b0
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/178838
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>