GOCBC-1159: Refresh DNS SRV when cluster uncontactable
Motivation
-----------
When DNS SRV records are used to connect to the SDK it's
possible for the underlying addresses to change, i.e. the
cluster to move. We need to detect this happening and react
accordingly so that the SDK can continue to work correctly.

Changes
-------
Add a callback to cccp poller for when no config can be fetched
from any node, once in steady state.
Add handler to agents to pick up when no config can be fetched
and refresh the DNS SRV record if applicable.
If the DNS SRV record is refreshed then check if it looks like a
new cluster location, if so then reset configs and push out new
fake configs with the new addresses.

Change-Id: I95d329c0f5f6b3bd93dfa838733e75f4ad3d6c97
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/179999
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>