GOCBC-1339: Readd CleanupWatchATRs to the transactions config
Motivation
----------
Whilst removing this field was deemed as acceptable it does cause
headaches with compatibility of gocb and gocbcore across versions.
We should readd the field but ignore its use and add comments
that make it clear that it should never be used.

Changes
-------
Readd CleanupWatchATRs field to the transactions config

Change-Id: I699aabf89e6d4d7eb31caae567efb8439c4d5e9d
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/180255
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>