GOCBC-1326: Add retry for when AuthProvider does not know a node
Motivation
----------
A scenario can arise where the cluster has changed but we haven't
yet received a config map to tell us about it and we request auth
credentials for a http request for a node which is rebalanced out.
When this happens tools using cbbs will no longer have any
credentials to give us. When this happens we should allow them to
inform us and we try another node.

Changes
-------
Add a new ErrNodeNotAvailable error.
Add a denylist for when selecting an address for use for a
http request.
Detect this error when returned from the AuthProvider Credentials
call during http requests and add the address to the denylist.
Add a new http node auth unknown retry reason.

Change-Id: I7765ea49f31deee77bcfc7fdcf0b211534dfe442
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/180554
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>