GOCBC-1354: Convert memd conn features to atomics
Motivation
----------
Now that we have added fetching a cluster config to bootstrap we
can trigger a race when the fetch fails. As fetching a cluster
config isn't fatal to bootstrap we can encounter the situation
where the server replies with the response after we have cancelled
the request our side. When this happen it's possible for enabling
features on the memd conn to race with read packet, which previously
was not possible as there could not be any requests in flight.

Changes
-------
Update memd conn features to use an atomic int rather than a map/

Change-Id: Ia0d4867171c56552c656b06ccba0485abf22f696
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/180769
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>