GOCBC-1322: Add functionality to wait until a config snapshot is available
Motivation
-----------
Sometimes users need to do things like take action based on how
many vbuckets there are in the bucket. In order to do this they
currently need to run a poll loop fetching the config snapshot,
until the snapshot revID is greater than -1.

Changes
-------
Add a WaitForConfigSnapshot function to agent and kvmux.
Add a config seen channel to kvmux which is closed when a config
with revID > -1 is seen.
Rework OnNewRouteConfig in kvmux slightly to be easier to
follow.

Change-Id: Ie6c07fe5f59cacd31942b1dbd427e24d02639490
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/182265
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>