GOCBC-1368: Update RangeScanCreate Range validation
Motivation
----------
We check that start and end are both set or that both exclusive
start and exclusive end are both set. This is incorrect, we
should check that either start or exclusive start are set and
that either end or exclusive end are set.

Changes
-------
Check that either start or exclusive start are set and that
either end or exclusive end are set.

Change-Id: Ibd98d04dae0540ee5d689f939bed2d5fcf733c70
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/183797
Reviewed-by: Graham Pople <graham.pople@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>