GOCBC-1387: Update memdclient close to wait for release channel
Motivation
----------
When we fail to  close the underling tcp connection used by a
memdclient we can trigger a race where the read side attempts a
read after we have released the connection (and reset internal state).
This can lead to a panic when the read attempts to access internal
buffers that are set to nil.

Changes
-------
Update memdclient close so that it shares logic with a successful
close - it will wait for the read side to stop before releasing
the connection.

Change-Id: I2756a70d3519ad9b3dcd645d80b4c1e15aeebe2d
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/186476
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>