GOCBC-1388: Allow multiple kv nodes to identify as seed in mux
Motivation
----------
If a node is 127.0.0.1 then ns server can send us a nodesExt entry
which has neither hostname or thisNode set. When this happens we
detect is the seed node even if it isn't. There is logic in the kv
muxer that assumes that only a single node in the cluster can be
identified as the seed node, and we end up in a situation
where the SDK tries to connect to an address with no hostname
or port.

Changes
-------
Update kvmux to allow multiple nodes marked as seed node to connect
over nonTLS.

Change-Id: Ie66139c5e8363e77c27749b71cd38c0a2b4a2089
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/186507
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>