GOCBC-1393: Invalidate n1ql cache entries on prepared failures
Motivation
-----------
When we execute a prepared statement and it fails with error codes
4040,4050,4060,4070,4080,4090 we should invalidate the cache entry
and attempt to re-prepare the statement.

Changes
--------
Refactor enhanced and old style prepared statement logic to share
more code.
Move retry handing for QueryPreparedStatementFailureRetryReason
to only run for prepared statements.
Do not retry the execution of already prepared statements when the
reason is QueryPreparedStatementFailureRetryReason, instead attempt
to reprepare the statement.
Invalidate the cache entry if the error is due to
QueryPreparedStatementFailureRetryReason

Change-Id: I0d1130a76e3b2646a6c65343255c009eee2eea42
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/187552
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>