GOCBC-1401: Expose SeqNo on DCP rollback error
Motivation
----------
When OpenStream responds with a rollback error the payload contains
a sequence number. This sequence number can be used by calling
applications to know where to restart their stream, rather than
rolling back to 0. We should expose this.

Changes
--------
Add a DCPRollbackError type containing the underlying error and
a SeqNo.
Check for ErrMemdRollback in the OpenStream callback, extract
SeqNo from the value and decorate the error with the new
DCPRollbackError.

Change-Id: I23050f1945afe9b57a918d3f8f125e67076e9bb9
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/189194
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>