GOCBC-1401: Expose SeqNo on DCP rollback error
Motivation
----------
When OpenStream responds with a rollback error the payload contains
a sequence number. This sequence number can be used by calling
applications to know where to restart their stream, rather than
rolling back to 0. We should expose this.

Changes
--------
Add a DCPRollbackError type containing the underlying error and
a SeqNo.
Check for ErrMemdRollback in the OpenStream callback, extract
SeqNo from the value and decorate the error with the new
DCPRollbackError.

Change-Id: Ia3ac75690a37463763ae0bc537a4352b29434098
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/189193
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>