GOCBC-1403: Update cccp poller to not wait for config before running
Motivation
----------
We pipelined fetching a config into bootstrap, and at that time
we also updated the cccp poller to wait for a connection to fetch
a config before starting polling. A config fetch failing in
bootstrap, or fetching an invalid config, is not a fail condition
for bootstrap. This means that if there is one seed endpoint
provided to the SDK and the config that is fetched as a part of its
bootstrap is invalid then cccp will never start. This ends up being
unrecoverable for the SDK.

Changes
-------
Update cccp poller to not wait util connections have fetched a config
before starting polling.
Add logic to kv muxer to not requeue get cluster config requests as
they are always sent directly to a specific pipeline.

Change-Id: Idad5f91d342429abb485dd89cb119fc460ee51b3
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/189520
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>