Always use snapshots in range scan tests
Motivation
----------
Our CI tests are flakey because the server thinks that there are
no relevant keys in the vbucket when we run our tests. We need
to use snapshot requirements to ensure that the server waits for
the data to be present.

Changes
-------
Update range scan tests to use snapshot.

Change-Id: I84d5398ad25eb5284dbbb1ac16fd4a9931dbd2a5
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/190234
Reviewed-by: Emilien Bevierre <emilien.bevierre@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>