GOCBC-1410: Return early if applying config to mux during shutdown
Motivation
----------
If there's a request in flight when we shutdown the kvmux then
it's possible for the response to come in with a NMVB. If this
happens after we've set the clientMux to nil but before we
drain the pipelines then we trigger a panic in OnNewRouteConfig.

Changes
-------
Return from OnNewRouteConfig early if client mux is nil.
Move waiting for poller shutdown to before we shutdown
the muxer.

Change-Id: Ia4efb33e172caea8a0b9e92193a5cefb290f8248
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/190766
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>