GOCBC-1410: Return early applying config to http mux during shutdown
Motivation
----------
We aleady made a change to pevent a config fom being applied to
the kv mux if the mux is shutting down. We need to do the same
in the http muxer. Whilst the http muxer does remove itself
from the config watchers as the first step of shutdown there is
still a possible race meaning that OnNewRouteConfig can be called
after the mux is cleared.

Changes
-------
Return early when applying a config to the http mux if the client
mux is nil.

Change-Id: I285bb9621eaf23ebfba1cd12bb9ddccf29b94c12
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/190796
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>