GOCBC-1415: Take lost cleanup locations pointer during close
Motivation
----------
We have a data race on transactions close if one of the atr
locations is deleted from the server at the same time. This is
because we don't actually copy the locations map during shutdown,
we only take a copy of the pointer.

Changes
--------
Make copy of locations map during shutdown.

Change-Id: Ic8aee4197f201456046c5e929d60eb80efbb13bb
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/191510
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>