GOCBC-1429: Update seed node config parsing
Motivation
-----------
When the SDk is put into ns_server mode it will not connect to
the "seed node" over TLS, but will connect to all other nodes using
TLS. In the case where we receive a NMV response in this mode it's
possible for us to parse the config incorrectly - identifying a remote
node as the seed node. This leads to us attempting to shutdown the
connection on which the NMV config came from. This leads to a
deadlock where memdClient close is waiting for the read thread
to complete, but close is being called from that thread. When
this happens it's also within the kv mux config lock, which prevents
any futher config updates from being applied.

Changes
--------
Update "ns_server" mode to only allow loopback addresses to be used.
When creating the first route config identify the seed node in the
config via the thisNode entry. Record both the hostname provided
by the user and a node identifier made of ${hostname}:${services.mgmt}.
On futher config updates identify the seed node using the node
identifier and swap out that hostname for the user provided hostname.

Change-Id: Ida4fcdb83bf0a207de94a389a1942f9902849854
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/192107
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/192506