GOCBC-1436: Check if readFromReplica supported when used
Motivation
----------
When use_replica is specified in a query payload we should first
check to see if we know that the feature is unsupported. If
unsupported then we should fast fail.

Changes
--------
Return ErrFeatureNotAvailable when use_replica set in the query
payload but n1ql.readFromReplica is not in cluster capabilities.

Change-Id: I333eb3d8d901ef799e89bc48b078767e401c061f
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/193182
Reviewed-by: Mateusz <matt.wozakowski@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>