GOCBC-1441: Revert GOCBC-1413 changes
Motivation
----------
The change that we made for GOCBC-1413 updated NoTLSSeedNode
being true to return an error if TLS was not also enabled.
This has caused issues and also makes it so that if tls was
disabled on create but is then enabled during reconfigure
security then NoTLSSeedNode can not be set.

Changes
--------
Revert GOCBC-1413 changes so that we do not return an error
if tls is disabled but NoTLSSeedNode is enabled.

This reverts commit 3b22127f88aa30791165f6585d70c96c5ad0af3a.

Change-Id: I4d01ba352c03944f07aee3f5f93df12455a6a311
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/193191
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>