GOCBC-1439: Add support for cluster config known versions
Motivation
----------
The server has added functionality to only send a config in
reponse to CCCP when the config is newer than we've already seen.

Changes
-------
Add support for the FeatureClusterMapKnownVersion hello feature
and always hello with it.
Update CCCP poller to be able to pull the current known revision
id and epoch from the config manager.
When revid/epoch are present and the feature is available send
our current known config as a part of the CCCP request.

Change-Id: I1ade6d38352fdc504a117a0c191b2719bd3a45a6
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/193389
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Tested-by: Charles Dixon <chvckd@gmail.com>