GOCBC-1439: Add support for brief cluster map notifications
Motivation
-----------
To reduce traffic and data volumes the server has added support
for brief cluster map notifications. These allow the server to
tell us when it has a newer config than it thinks we already has,
and removes the need for us to poll.

Changes
-------
Add CmdMagicServerReq for requests that originate from the server.
Add a server request handler to kvmux and pipe it into memdclient,
to be called when the packet magic is server request - before we
attempt to match the opaque to our own request.
Add a handler for brief cluster map notifications to cfg manager,
this will send a get cluster config request if the notification
is newer than the config we have - this is debounced so that only
a single request is ever in flight at once.
Update the cccp poller to skips nodes which support brief cluster
map notifications.

Change-Id: I31ca2a77e58c8fd257b2c627d845a8a64b4beb66
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/194011
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>