GOCBC-1453: Stop http poller inside poller controller lock
Motivation
----------
As a part of faster failover changes we simplified the pollers so
that Stop will wait until the pollers have fully completed. When
the poller controller receives a new config and http polling is use
we stop the http poller but previously we were not waiting for it
to fully stop. Now that we do there is a race on an internal channel.

Changes
-------
Call http poller stop within the poller controller lock when stopping
the poller on new config.

Change-Id: I70bbd86634977358cebeb48c645a165b8d079559
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/194402
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>