GOCBC-1457: Reduce default http idle timeout to 1s
Motivation
----------
The current SDK 4.5s HTTP idle timeout is too close to the server
timeout so we can have cases where a request is made but an EOF
is received because the request is in flight when the connection
is closed by the server.

Changes
-------
Update agents to use default http idle timeout of 1s.

Change-Id: I9e196d9ba525b23edbaeaa210b047cd76f6f8955
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/195378
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>