GOCBC-1466: Always close CCCP done channel on exit
Motivation
-----------
When the user specifies a non-existent bucket or a memcached
bucket there is an edge case race which can lead to a deadlock.
The bootstrap process tells the poller controller to stop the
cccp poller. The cccp poller has managed to send and receive its
on request and has come to the same conclusion. Before Stop is
called on cccp poller the poller has already existed. This can
lead to Stop being called on a poller which has exiting without
closing the done channel, which then currently waits on a channel
that will not be closed.

Changes
-------
Always close the done channel on CCCP poller exit.

Change-Id: Ibd6f4b69c79c5759702590c3beb4caed80cbfe72
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/195806
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>