GOCBC-1468: Skip sending current rev for get config if < 1
Motivation
----------
There's an edge case where bootstrap can fetch an invalid bucket
config but the server then tells us that it has a config we should
fetch. When this happens we haven't yet seen a config and we write
-1 into the extras. The command does not allow revids of <1 so we
get an invalid request back.

Changes
-------
Update cccpConfigFetcher to only write the rev information to
the extras if the revID > 0.

Change-Id: I382ebb28c35548873f74e30cc705683b1564b42a
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/195904
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>