GOCBC-1451: Disable cccp polling for DCP agent by default
Motivation
----------
A DCP connection (with streams established) will automatically
tell you about any changes in vBucket state on that node - if a
vbucket changes from the state it was in when the stream was
established; the stream will be ended via a
StreamEnd::StateChanged message.  In other words, if you're only
using DCP, then GET_CLUSTER_CONFIG on the DCP connection(s) are
of little value; and moreover are at risk of high runtimes (from
client's pov) as they must drain all the outstanding DCP messages
before they will see the GET_CLUSTER_CONFIG response.

When topology changes the stream will be ended, at which point
the user will send a new stream open request and if the cluster
topology has indeed changed (and no CCCP request was sent in the
interim) then if the request went to the wrong node the SDK would
get a not my vbucket error at which point it'd update its
topology anyway.

Due to this we should only use HTTP polling for DCP.

Changes
-------
Refactor poller controller Run function.
Update dcp agent to only run http polling by default.
Add a field to DCPConfig to allow users to enable CCCP polling.

Change-Id: I5834a2ab83ddbdfd9d6778b148736d99affc4b6b
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/195981
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>