GOCBC-1469: Only set bucket capabilities when received bucket config
Motivation
----------
When we create a new bucket and immediately connect to it then
it's possible for select bucket to fail. When this happens it's
still possible for us to successfully get and apply a cluster
config, which will not contain any bucket settings. If the agent
is bucketful then we should not set bucket settings to unsupported
until we receive an actual bucket config.

Changes
-------
Update kvmuxstate to be aware of whether configs it receives should
have a bucket.
Do not set bucket capability statuses until we receive a config
for the specified bucket.

Change-Id: I38b5dcda1266f4312ac6bdf5e4ebc8e777806e4a
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/196376
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>