GOCBC-1225: Update TestRandomGet
Motivation
----------
TestRandomGet sporadically fails on CI due to the fetched document
having no value. We create documents with no value in other tests
so get random is likely picking up one of these docs.

Changes
-------
Update TestRandomGet to create a new collection and then use that
collection for the test.
If collections are not supported then do not check for a value.

Change-Id: Iacd2b0997b6752e0ff29dcb502759b4ea4800ff6
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/196382
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>