GOCBC-1471: Rework how pollers are stopped
Motivation
----------
When an agent is created with only a cccp or a http poller then we
can encounter a hang on agent Close if it is called immediately
after Connect. This is because the poller controller Stop is called
because it is fully running - the controller calls Stop on the
underlying poller before it is started, and then the controller
exits the run loop before starting the poller.

Changes
-------
Remove the done channels from the pollers, and do not block in the
Stop function - instead allowing the return of DoLoop to signal that
the poller has finished.
Update poller controller to have its own stop channel which is closed
once Stop has been called and the poller DoLoop has completed.

Change-Id: I4e36f4be53d96855c172d9a30c547f5db4c70828
Reviewed-on: https://review.couchbase.org/c/gocbcore/+/196652
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>